// Web_IDE_Proxy_Server/src/types/db.ts
export interface Project {
    id: string;
    user_id: string;
    name: string;
    template: string;
    language: string;
    visibility: "private" | "public";
    created_at: string;
    updated_at: string;
}

export type FileNode = {
    id: string
    project_id: string
    parent_id: string | null
    name: string
    type: "file" | "folder"
    content: string | null
    created_at: string;
    updated_at: string;
}